Removes 1 un-needed interation
Since document.getElementsByTagName() will return an array of
elements starting with 0 and length will give the number of
items. The loop in fallbacks() will start iterate over an
element in this.images that doesn't exists.

By changing -- operator to a prefix instead of a suffix
it will now start on an element that actually exists.